requirements.txt
python-binance==1.0.19
flask==2.3.3
flask-socketio==5.3.6
pandas==2.0.3
numpy==1.24.3
TA-Lib
python-engineio==4.7.1
python-socketio==5.8.0
websocket-client==1.6.3
redis==4.5.5
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.2
